etf-dash/requirements.txt
ansi2html==1.8.0
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.9
boto==2.49.0
boto3==1.26.160
botocore==1.29.160
Brotli==1.0.9
bz2file==0.98
catalogue==2.0.8
catboost==1.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
dash==2.11.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docutils==0.20.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
Flask-Compress==1.13
fonttools==4.40.0
fqdn==1.5.1
future==0.18.3
gensim==4.3.1
graphviz==0.20.1
greenlet==2.0.2
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
JPype1==1.4.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
konlpy==0.6.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook_shim==0.2.3
numpy==1.25.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyMySQL==1.0.3
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.17
srsly==2.4.6
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
wordcloud==1.9.2
